Add testvectors hashes in META.yml
This includes the known-answer test hashes from
https://github.com/pq-crystals/dilithium/tree/master merged into a single
META.yml file.
It also includes alternative testvector hashes from PQClean
https://github.com/PQClean/PQClean/tree
(those also test the detached signature API).

Additionally, we copy over a script from mlkem-native that allows querying it.
It either uses yq to properly parse the yml, or if that is not available, it
falls back to string processing grep which works fine for the simple structure.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>