fix memory error
The buffer for the message passed to crypto_sign_open needs to be
mlen + CRYPTO_BYTES bytes according to the NIST API. If signature
verification fails, this gets set to 0 entirely. In one place
the buffer size used was only mlen which caused a seg fault in
gcc4.8.

We should use asan to catch mistakes like this earlier in the future

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>