CBMC: Add contract and proof for poly_pointwise_montgomery
Resolves https://github.com/pq-code-package/mldsa-native/issues/107

This also adjusts the bounds of montgomery_reduce. If the output bounds are not
required (as in this case), the inputs can be much larger.
The maximum value that is relevant for Montgomery multiplication is INT32_MIN^2.
This way we can prove poly_pointwise_montgomery without placing any requirements
on the sizes of the inputs.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>