Replace arithmetic preprocessor conditions by MLDSA_MODE
There are multiple places in the code base where a parameter-set
preprocessor branch is conducted not by checking MLDSA_MODE but
the specific value of one of the derived constants, such as
MLDSA_GAMMA1 or MLDSA_GAMMA2.

This commit replaces those branch conditions by equality check
for MLDSA_MODE. It also reorders the branches so that MLDSA_MODE==2
comes first.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>