CI: Add dependabot
The mldsa-native CI uses quite a few external dependencies (such as
nix-community/cache-nix-action, aws-actions/configure-aws-credentials, etc.).
We should keep those up to date to avoid problems in the long term.
This commit adds a dependabot configuration that will check for new versions
once a month and opens PRs in case there are updates.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>