CBMC: Bound nblocks by 8 in {keccak,shake128,shake256}_squeezeblocks
{keccak,shake128,shake256}_squeezeblocks take an arbitrary number of
blocks that are to be squeezed. The previous contracts used UINT32_MAX/rate
as an upper limit. More relistically it should probably be a SIZE_MAX/rate,
but as we are never using them with large values either is fine.

Unfortunately, the existing bound causes the proofs to fail in CI (likely due
to excessive memory consumption).

This commit instead bounds the maximum number of blocks much more aggressively
to be at most 8. This is likely more than we need in all call sites in
ML-DSA and it results in passing proofs in CI.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>