Debug: Add bounds assertions macros
Following mlkem-native, this commit adds debug.h and debug.c,
offering assertions for runtime checking of array bounds.
Those assertions provide the main source of confidence in the
validity of our bounds assumptions prior to completion of all
CBMC proofs up to the toplevel API.

All implementations of functions in poly.c have bounds assertions
for inputs and outputs added that match the CBMC specifications.

As for mlkem-native, if CBMC is enabled, the bounds assertions
are interpreted as CBMC proof obligations.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>