Refactor stream256 code to using xof256 naming
The poly_uniform_gamma1 function is the last one using the stream256
instead of the standard keccak APIs.
This commit refactors it to be in-line with the batched implementation as
well as the xof128 function.

The initialization function is removed and inlined into poly_uniform_gamma1.

The CBMC proof is adjusted accordingly.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>