Add explicit _release function call for xof128/xof256
Some Keccak implementations (e.g., liboqs) may dynamically allocate the Keccak
state in the _init() function. We, hence, need an explicit _release() function
to potentially free this memory.
In mldsa-native this is a no-op (or it zeros out the state once we have
secure stack zeroziation).

This commit adds the _release functions for both shake128 and shake256.
According mld_xof{128,256}_release macros are added to symmetric.h.
Function calls to the _release functions are added in all places that use the
mld_xof APIs.

CBMC proofs are amended for those functions to in-line the empty release
functions.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>