Fix: conditional in nix-post-check step in setup-nix action
The nix-post-check step in the setup-nix action used a conditional that is incorrectly using ${{ steps.nix-pre-check.outputs.installed }} as a raw value instead of a string causing a error in case the value is undefined.
This commit wraps it quotes fixing the issue.

Related to https://github.com/pq-code-package/mlkem-native/issues/1033

Signed-off-by: willieyz <willie.zhao@chelpis.com>