Refactor polyvecl_pointwise_acc_montgomery in preparation of native code
This commit refactors polyvecl_pointwise_acc_montgomery in preparation of
pulling in native code for this function.

For the optimized backends, this function will perform the accumulation of the
products on 64-bit words prior to reducing it back to 32-bit, e.g., using
Montgomery reduction.

This requires a pre-condition that this accumulation cannot overflow an
int64_t. This is the case as in all call sites the second input is
the output of an NTT which is bounded by 9q.

As a side-effect this commit allows proving the bounds using CBMC.
Before the function was simply calling poly_pointwise_montgomery
which does not provide any bounds on the inputs/outputs.
Consequently the proof of the accumulation of those outputs to not overflow
would be impossible.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>