CI: Refactor to align with mlkem-native
This commit refactors our CI to split it in different workflows:
Base, Extended, CBMC, Nix.

This way we can first run the base CI before invoking the other workflows.
It also allows the Extended and CBMC workflows to wait for the nix cache
to be built in case it needs to change.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>