AArch64: Add native implementation of rej_eta
- Implements mld_rej_uniform_eta2_aarch64() for eta=2 case
- Implements mld_rej_uniform_eta4_aarch64() for eta=4 case
- Autogenerates required look-up table
- Resolves #319

Note that we reject modulo a very small number (9 or 15). Doing that in 32-bits
appears wasteful. Doing it on 8-bit values makes the look-up table approach
infeasible as it would require a 2^16-entry table. We insetead opt for the
16-bit approch which is close to the ML-KEM rejection sampling.
Some additional arithmetic is required which we merge into the final
copying loop. That is also where the expansion from 16-bit values to
32-bit values happens.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>