CI: Test config variation optblocker (i.e., no inline asm barrier)
This commit adds a config variation test to CI to test the alternative
constant time functions (when no inline asm is available).
To do so, first a MLD_CONFIG_FILE configuration option is introduced
allowing to define a custom configurtion file.

This option is then used in a workflow to include a config that has
MLD_CONFIG_NO_ASM set.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>