CT: Make polyvecl_chknorm and polyveck_chknorm constant flow
The reference implementation implements polyvecl_chknorm and polyveck_chknorm
in variable time, i.e., it leaks which polynomial violated the rejection bound.

This approach appears safe, but somewhat unclean as it is still operating on
_secret data_. When performing constant-time testing it also requires a
number of declassifications.

This commit takes a more conservative approach and changes both funcitons to
constant-time implementations in the hope tha the performance penalty is
acceptable.

A minor change is that the API is changed to returning
0xFFFFFFFF in the case of failure to be able to re-use existing constant-time
primitives.
CBMC proofs are adjusted accordingly.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>