Adjust KAT test to work with PCT
Currently our KAT tests do not work in case the pairwise-consistency
test (PCT) is enabled through MLD_CONFIG_KEYGEN_PCT.
The reason for this is that when PCT is enabled, a signature is generated
at the end of key generation which requires randomness to be sampled
from randombytes() messing with the state of the deterministic
random number generator used for KAT tests.

To work around this, this commit changes the KAT tests to follow the same
approach that is also used by mlkem-native: Only use the de-randomized APIs
and derive all random coins deterministically using SHAKE instead of
randombytes().

This enabled KAT tests even when PCT is enabled.
Unfortunately, it requires changing the KAT hashes once.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>